# CUDA: faster tile FA (Pascal/AMD), headsize 256