This PR refactors and deduplicates the CUDA "tile" FlashAttention kernels, adds support for head size 256, and improves performance for Pascal/AMD without FP16 mma. If fast FP16 math is available it is now used, but the KQ accumulation and the softmax are always down with FP32 precision (as these seem to be the numerically problematic parts of the kernel). The kernel now has a more flexible parameterization, I tuned `kq_stride` and `kq_nbatch` for P40/RX 6800/Mi 50. It is also possible to use a `warp_size` of 64 rather than 32 but I was not able to get better performance this way; I'm keeping the functionality since it's possible I'm currently ignorant about something and may want to revisit this in the future.

<details>
<summary>Performance changes</summary>

| GPU           | Model           |     Microbatch size | Test      |     t/s master |     t/s c9a318a5a |     Speedup |
| :------------ | :-------------- | ------------------: | :-------- | -------------: | ----------------: | ----------: |
| MI50          | llama 1B Q4_0   |                  16 | pp16384   |         188.69 |            682.28 |        3.62 |
| MI50          | llama 1B Q4_0   |                  32 | pp16384   |         184.60 |            956.19 |        5.18 |
| MI50          | llama 1B Q4_0   |                 512 | pp16384   |         185.61 |           1510.04 |        8.14 |
| MI50          | llama 8B Q4_0   |                  16 | pp16384   |          37.08 |            194.09 |        5.23 |
| MI50          | llama 8B Q4_0   |                  32 | pp16384   |          38.41 |            163.29 |        4.25 |
| MI50          | llama 8B Q4_0   |                 512 | pp16384   |          38.35 |            203.44 |        5.31 |
| RX 6800       | llama 1B Q4_0   |                  16 | pp16384   |         378.27 |            645.38 |        1.71 |
| RX 6800       | llama 1B Q4_0   |                  32 | pp16384   |         275.62 |            659.44 |        2.39 |
| RX 6800       | llama 1B Q4_0   |                 512 | pp16384   |         332.24 |            898.23 |        2.70 |
| RX 6800       | llama 8B Q4_0   |                  16 | pp16384   |          55.59 |            172.44 |        3.10 |
| RX 6800       | llama 8B Q4_0   |                  32 | pp16384   |          71.36 |            174.42 |        2.44 |
| RX 6800       | llama 8B Q4_0   |                 512 | pp16384   |          84.39 |            231.21 |        2.74 |
| P40           | llama 1B Q4_0   |                  16 | pp16384   |        1109.57 |           1189.42 |        1.07 |
| P40           | llama 1B Q4_0   |                  32 | pp16384   |        1500.00 |           1686.71 |        1.12 |
| P40           | llama 1B Q4_0   |                 512 | pp16384   |        2052.54 |           2449.76 |        1.19 |
| P40           | llama 8B Q4_0   |                  16 | pp16384   |         279.17 |            296.43 |        1.06 |
| P40           | llama 8B Q4_0   |                  32 | pp16384   |         357.32 |            356.78 |        1.00 |
| P40           | llama 8B Q4_0   |                 512 | pp16384   |         497.47 |            499.94 |        1.00 |

</details>