@JohannesGaessler 2025-09-07
The support for the combination of head size 256 + quantized KV cache still has other issues and requires a refactor of the "vector" kernels.